Refactor TD tests to own folder
[ghstack-poisoned]
