+++ tools/test/heuristics/test_heuristics.py
@@ -3,7 +3,7 @@ import json
 import pathlib
 import sys
 import unittest
-from typing import Any, Dict, Optional, Set
+from typing import Any, Dict, Set
 from unittest import mock
 
 REPO_ROOT = pathlib.Path(__file__).resolve().parent.parent.parent
@@ -11,6 +11,7 @@ try:
     # using tools/ to optimize test run.
     sys.path.append(str(REPO_ROOT))
 
+    from tools.test.heuristics.heuristics_test_mixin import HeuristicsTestMixin
     from tools.testing.target_determination.determinator import (
         AggregatedHeuristics,
         get_test_prioritizations,
@@ -20,7 +21,7 @@ try:
     from tools.testing.target_determination.heuristics.previously_failed_in_pr import (
         _get_previously_failing_tests,
     )
-    from tools.testing.test_run import TestRun, TestRuns
+    from tools.testing.test_run import TestRun
 
 except ModuleNotFoundError:
     print("Can't import required modules, exiting")
@@ -32,36 +33,6 @@ def mocked_file(contents: Dict[Any, Any]) -> io.IOBase:
     json.dump(contents, file_object)
     file_object.seek(0)
     return file_object
-
-
-class HeuristicsTestMixin(unittest.TestCase):
-    def assert_heuristics_match(
-        self,
-        test_prioritizations: TestPrioritizations,
-        expected_high_tests: Optional[TestRuns] = None,
-        expected_probable_tests: Optional[TestRuns] = None,
-        expected_unranked_tests: Optional[TestRuns] = None,
-    ) -> None:
-        if expected_unranked_tests:
-            self.assertTupleEqual(
-                test_prioritizations.get_unranked_relevance_tests(),
-                expected_unranked_tests,
-                "Unranked tests differ",
-            )
-
-        if expected_probable_tests:
-            self.assertTupleEqual(
-                test_prioritizations.get_probable_relevance_tests(),
-                expected_probable_tests,
-                "Probable relevance tests differ",
-            )
-
-        if expected_high_tests:
-            self.assertTupleEqual(
-                test_prioritizations.get_high_relevance_tests(),
-                expected_high_tests,
-                "High relevance tests differ",
-            )
 
 
 class TestParsePrevTests(HeuristicsTestMixin):
@@ -127,11 +98,8 @@ class TestParsePrevTests(HeuristicsTestMixin):
             tests
         ).get_aggregated_priorities()
 
-        self.assert_heuristics_match(
-            test_prioritizations,
-            expected_high_tests=expected_prioritizations.get_high_relevance_tests(),
-            expected_probable_tests=expected_prioritizations.get_probable_relevance_tests(),
-            expected_unranked_tests=expected_prioritizations.get_unranked_relevance_tests(),
+        self.assertHeuristicsMatch(
+            test_prioritizations, expected_prioritizations=expected_prioritizations
         )
 
 
@@ -154,7 +122,7 @@ class TestInterface(HeuristicsTestMixin):
             TestRun("test5"),
         )
 
-        self.assert_heuristics_match(
+        self.assertHeuristicsMatch(
             prioritizations,
             expected_probable_tests=expected_probable_tests,
             expected_unranked_tests=expected_unranked_tests,
@@ -197,7 +165,7 @@ class TestAggregatedHeuristics(HeuristicsTestMixin):
         print("Aggregated Heuristics")
         aggregated_pris = aggregator.get_aggregated_priorities()
 
-        self.assert_heuristics_match(
+        self.assertHeuristicsMatch(
             aggregated_pris,
             expected_high_tests=expected_high_relevance,
             expected_probable_tests=expected_probable_relevance,
@@ -234,7 +202,7 @@ class TestAggregatedHeuristics(HeuristicsTestMixin):
 
         aggregated_pris = aggregator.get_aggregated_priorities()
 
-        self.assert_heuristics_match(
+        self.assertHeuristicsMatch(
             aggregated_pris,
             expected_high_tests=expected_aggregated_high_relevance,
             expected_probable_tests=expected_aggregated_probable_relevance,
